pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sclip"
version = "0.2.0"
description = "SmartClip AI - Transform long-form videos into viral-ready short clips using AI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Sarian"}
]
keywords = [
    "video",
    "clips",
    "ai",
    "whisper",
    "groq",
    "gemini",
    "youtube",
    "tiktok",
    "reels",
    "shorts",
    "captions",
    "ffmpeg",
    "transcription"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Video :: Conversion",
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "groq>=0.4.0",
    "yt-dlp>=2024.1.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
gemini = [
    "google-genai>=1.0.0",
]
openai = [
    "openai>=1.0.0",
]
local = [
    "faster-whisper>=1.0.0",
]
all = [
    "google-genai>=1.0.0",
    "openai>=1.0.0",
    "faster-whisper>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
sclip = "src.main:main"

[project.urls]
Homepage = "https://github.com/sarian/sclip"
Repository = "https://github.com/sarian/sclip"
Issues = "https://github.com/sarian/sclip/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
    | __pycache__
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
